Added logic to login page of admin used session to store the login details of admin so admin dont need to log in often

+++ Student_registration_portal/src/main/java/com/controller/RegistrationController.java
@@ -6,8 +6,10 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.util.List;
 
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
-
+import com.service.StudentServiceImpls;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -26,15 +28,21 @@ import com.service.StudentService;
 @Controller
 public class RegistrationController {
 
+    private final StudentServiceImpls studentServiceImpls;
+
 	@Autowired
 	StudentService studentService;
+
+    RegistrationController(StudentServiceImpls studentServiceImpls) {
+        this.studentServiceImpls = studentServiceImpls;
+    }
 	@RequestMapping("/home")
 	String index() {
 		return "home";
 	}
 		
 	@PostMapping("/save")
-	String save(@ModelAttribute StudentModel studentModel,Model model) {
+	String save(@ModelAttribute StudentModel studentModel,Model model,HttpServletRequest request) {
 		
 		System.out.println(studentModel);
 		System.out.println(studentModel.getId());
@@ -46,22 +54,35 @@ public class RegistrationController {
 		}
 		boolean bool=studentService.saveStudent(studentModel);
 		if(bool && studentModel.getId()!=0) {
-			System.out.println("edit");
+//			System.out.println("edit");
+			model.addAttribute("url","admin");
 			model.addAttribute("message","Detail Updated Successfull..");
 		}
 		else if(bool) {
-			System.out.println("save");
+//			System.out.println("save");
 			model.addAttribute("message","Registration Successfull..");
 		} else {
 			model.addAttribute("message","Failed..");
 		}
+		
+//		String arr[]=request.getHeader("Referer").split("/");
+//		if(arr[arr.length-2]=="edit")
+//			model.addAttribute("url",request.getHeader("Referer"));
+//		else model.addAttribute("url",null);
+//		System.out.println(arr[arr.length-2]);
 		return "success";
 	}
 	    
 	@RequestMapping("/delete/{id}")
-	String delete(@PathVariable("id") int id) {
-		studentService.deleteStudent(id);
-		return "redirect:/admin";
+	String delete(@PathVariable("id") int id,HttpSession session) {
+		
+		if(session.getAttribute("adminEmail")!=null) {
+			studentService.deleteStudent(id);
+			return "redirect:/adminValidate";
+			}else {
+				return "redirect:/adminLogin";
+			}
+		
 	}
 	
 	@RequestMapping("/adminLogin")
@@ -71,11 +92,25 @@ public class RegistrationController {
 	}
 	
 	@RequestMapping("/adminPage")
-	String AdminPage(Model model) {
-		List<StudentModel>st= studentService.getStudentDetails();
-//		System.out.println(st.get(0).getBase64Image()+"\n"+"hiie");
-		model.addAttribute("students",st);
-		return "adminpage";
+	String AdminPage(Model model,@RequestParam("email") String email,@RequestParam("password") String password,HttpSession session) {
+//		System.out.println(email);
+//		System.out.println(email.equals("admin@gmail.com"));
+//		System.out.println(password);
+//		System.out.println(password.equals("admin"));
+//		System.out.println(request.getHeader("referer"));
+		if(email.equals("admin@gmail.com") && password.equals("admin")) {
+			List<StudentModel>st= studentService.getStudentDetails();
+//			System.out.println(st.get(0).getBase64Image()+"\n"+"hiie");
+			model.addAttribute("students",st);
+			session.setMaxInactiveInterval(900);
+			session.setAttribute("adminEmail",email);
+			return "redirect:/adminValidate";
+		}
+		else {
+			model.addAttribute("message","Login Failed.."); 
+			model.addAttribute("reasone", "Invalid Credentials.!");
+			return "success";
+		}
 	}
 	
 	@RequestMapping("/studentLogin")
@@ -84,18 +119,41 @@ public class RegistrationController {
 		return "studentLogin";
 	}
 	@RequestMapping("/edit/{id}")
-	String edit(@PathVariable("id") int id,Model model)
+	String edit(@PathVariable("id") int id,Model model,HttpSession session)
 	{
+		if(session.getAttribute("adminEmail")!=null) {
 		StudentModel student =studentService.getSingleStudentDetails(id);
 		model.addAttribute("student",student);
-		
 		return "editForm";
+		}else {
+			return "redirect:/adminLogin";
+		}
+		
 	}
 	
 	@RequestMapping("/studentRegistration")
 	String register()
 	{
 		return "studentReistration";
+	}
+	
+	@RequestMapping("/adminValidate")
+	String adminValidate(Model model,HttpSession session)
+	{
+		if(session.getAttribute("adminEmail")!=null) {
+		List<StudentModel>st= studentService.getStudentDetails();
+		model.addAttribute("students",st);
+		return "adminpage";
+		}else {
+			return "redirect:/adminLogin";
+		}
+	}
+	
+	@RequestMapping("/logout")
+	String logout(HttpSession session)
+	{	
+			session.invalidate();
+			return "redirect:/home";
 	}
